fix(storage) reusable configuration table component, improve naming
