+++ src/pages/storage/forms/StorageConfigurationRow.tsx
@@ -2,12 +2,15 @@ import { ReactElement, ReactNode } from "react";
 import { CpuLimit, MemoryLimit } from "types/limits";
 import { MainTableRow } from "@canonical/react-components/dist/components/MainTable/MainTable";
 import { FormikProps } from "formik/dist/types";
-import { NetworkFormValues } from "pages/networks/forms/NetworkForm";
-import { getLxdDefault } from "util/networks";
+import { getLxdDefault } from "util/storageVolume";
+import { StorageVolumeFormValues } from "pages/storage/forms/StorageVolumeForm";
+import { useQuery } from "@tanstack/react-query";
+import { queryKeys } from "util/queryKeys";
+import { fetchStoragePool } from "api/storage-pools";
 import { getConfigurationRow } from "components/ConfigurationRow";
 
 interface Props {
-  formik: FormikProps<NetworkFormValues>;
+  formik: FormikProps<StorageVolumeFormValues>;
   name: string;
   label: string | ReactNode;
   children: ReactElement;
@@ -16,7 +19,7 @@ interface Props {
   help?: string;
 }
 
-export const getConfigurationRowNetwork = ({
+export const getStorageConfigurationRow = ({
   formik,
   name,
   label,
@@ -28,8 +31,15 @@ export const getConfigurationRowNetwork = ({
   const values = formik.values as unknown as Record<string, string | undefined>;
   const isOverridden = values[name] !== undefined;
 
-  const [inheritedValue, inheritSource] = [getLxdDefault(name), "LXD"];
-  const isReadOnly = formik.values.readOnly;
+  // when creating the defaults can be set on the storage pool
+  const { data: storagePool } = useQuery({
+    queryKey: [queryKeys.storage, formik.values.pool, formik.values.project],
+    queryFn: () => fetchStoragePool(formik.values.pool, formik.values.project),
+    enabled: formik.values.isCreating,
+  });
+
+  const [inheritedValue, inheritSource] = getLxdDefault(name, storagePool);
+  const isReadOnly = formik.values.isReadOnly;
 
   const getOverrideValue = (): ReactNode => {
     if (typeof values[name] === "boolean") {
