+++ src/pages/storage/forms/StorageVolumeFormZFS.tsx
@@ -2,8 +2,8 @@ import React, { FC } from "react";
 import { CheckboxInput, Select } from "@canonical/react-components";
 import { StorageVolumeFormValues } from "pages/storage/forms/StorageVolumeForm";
 import { FormikProps } from "formik/dist/types";
-import ConfigurationTable from "pages/storage/forms/ConfigurationTable";
-import { getConfigurationRowStorage } from "pages/storage/forms/ConfigurationRowStorage";
+import ConfigurationTable from "components/ConfigurationTable";
+import { getStorageConfigurationRow } from "pages/storage/forms/StorageConfigurationRow";
 
 interface Props {
   formik: FormikProps<StorageVolumeFormValues>;
@@ -13,7 +13,7 @@ const StorageVolumeFormZFS: FC<Props> = ({ formik }) => {
   return (
     <ConfigurationTable
       rows={[
-        getConfigurationRowStorage({
+        getStorageConfigurationRow({
           formik: formik,
           label: "ZFS blocksize",
           name: "zfs_blocksize",
@@ -55,7 +55,7 @@ const StorageVolumeFormZFS: FC<Props> = ({ formik }) => {
           ),
         }),
 
-        getConfigurationRowStorage({
+        getStorageConfigurationRow({
           formik: formik,
           label: "ZFS block mode",
           name: "zfs_block_mode",
@@ -69,7 +69,7 @@ const StorageVolumeFormZFS: FC<Props> = ({ formik }) => {
           ),
         }),
 
-        getConfigurationRowStorage({
+        getStorageConfigurationRow({
           formik: formik,
           label: "ZFS delegate",
           name: "zfs_delegate",
@@ -83,7 +83,7 @@ const StorageVolumeFormZFS: FC<Props> = ({ formik }) => {
           ),
         }),
 
-        getConfigurationRowStorage({
+        getStorageConfigurationRow({
           formik: formik,
           label: "ZFS remove snapshots",
           name: "zfs_remove_snapshots",
@@ -97,7 +97,7 @@ const StorageVolumeFormZFS: FC<Props> = ({ formik }) => {
           ),
         }),
 
-        getConfigurationRowStorage({
+        getStorageConfigurationRow({
           formik: formik,
           label: "ZFS use refquota",
           name: "zfs_use_refquota",
@@ -111,7 +111,7 @@ const StorageVolumeFormZFS: FC<Props> = ({ formik }) => {
           ),
         }),
 
-        getConfigurationRowStorage({
+        getStorageConfigurationRow({
           formik: formik,
           label: "ZFS reserve space",
           name: "zfs_reserve_space",
